Rename data/ package to schemas/
